fix(bim): apply wasm locate + single-thread to all web-ifc init paths
Greptile review of #1465:

- setIfcWasmLocateFile previously only affected the writer, so its
  documented coverage of fromIfc was false: SpfReader (and the validation
  helpers) called api.Init() bare, so IFC import in a worker would still
  fail to find the wasm and would load the pthread build. Centralize web-ifc
  init in a shared `initIfcApi(api)` (locate handler + single-threaded) used
  by the writer, reader, and validation alike — making the public contract
  true and fixing the latent import/validation breakage in workers.
- Use the application/x-ifc MIME type for the IFC download instead of
  application/x-step.
